pyproject.toml
[tool.poetry]
name = "converciones"
version = "0.1.0"
description = ""
authors = ["isa <isaram1218@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyleniumio = "^1.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
